Updated top beta 1.6.2-R0.1

+++ ExpCraftDexterity/src/main/java/li/cryx/expcraft/dexterity/DexterityEntityListener.java
@@ -75,7 +75,7 @@ public class DexterityEntityListener implements Listener {
 				* 10.0 / plugin.getLevelCap() //
 				* plugin.getConfig().getDouble("Settings.FallDmgMultiplier"));
 		// apply reduction
-		int dmg = Math.max(0, event.getDamage() - damageReduction);
+		double dmg = Math.max(0, event.getDamage() - damageReduction);
 		event.setDamage(dmg);
 	}
